+++ spec/dummy/db/migrate/20260309000003_add_document_id_to_block_editor_threads.rb
@@ -1,5 +1,0 @@
-class AddDocumentIdToBlockEditorThreads < ActiveRecord::Migration[8.0]
-  def change
-    add_reference :block_editor_threads, :document, null: true, foreign_key: true
-  end
-end